src/message_iter: Maintain reference to underlying DBusMessage
Weirdly, an DBusMessageIter doesn't increase the refcount of the
referred DBusMessage, thus we need to maintain the reference manually to
prevent underlying DBusMessage from garbage collection before the
DBusMessageIter is collected, which results in a dangling reference.

Let's introduce a new type, lDBusMessageIter, which contains both the
iterator and the underlying DBusMessage, and refer/unref the message
during creation, initialization and garbage collection of the iterator.

Closes: https://github.com/daurnimator/ldbus/issues/20
Signed-off-by: Yao Zi <ziyao@disroot.org>